requirements.txt

joblib==0.15.1
Markdown==3.2.2
matplotlib==2.2.5
numpy==1.16.6
opencv-python==4.2.0.34
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3